Add 3264x1840 as a valid camera resolution for benefit of HTC DLX and others.
Currently, 3264x1836 is listed as a valid camera resolution option, but
3264x1840 is not.  The latter is what is exposed by the driver for the 6M 16:9
camera size on the HTC DLX (Droid DNA), not the 1836 option.  This is likely
true for other cameras as well, I would expect.  Without 3264x1840, the
highest resolution that is offered in a 16:9 format is 4M, far less than
the sensor is capable of.  3264x1840 is also correct for this device's sensor
when compared to the stock HTC ROM -- it takes photos at this size.

3264x1836 remains untouched for any devices that use it.  Given that these
values are checked against what the driver supports, and it's unlikely two
such similar resolutions would be supported by the driver, there shouldn't be
a duplicate.  I reused the existing text string resource for the 6M label, so
no new translations are needed.

Note that if testing this on a Droid DNA (or other device with the same
issue), you might see a crash when you select the new 6M option.  This is
an unrelated existing bug that occurs the first time you select any
widescreen size on this device.  It happens on the current nightlies as
well when selecting the 4M option.  It's not related to this change.
After the crash, the camera app can be reopened, the selected resolution is
still selected, and everything seems to work fine.  I'll look into this bug
at some point if I can finda solution, but it appears to affect existing
16:9 resolutions as well and thus shouldn't be a blocker for this change.

Change-Id: Ia75e46ff2cda12c0c3cbca58222c801b71f31115